# openapi.yaml
openapi: 3.1.0
info:
  title: Ego Microservices
  version: 1.0.0
servers:
  - url: http://${APISIX_HOST}:${APISIX_PORT}
paths:
  /auth/registration:
    post:
      tags:
        - auth
      summary: User Registration
      description: Регистрация пользователя
      operationId: user_registration_auth_registration_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensResponse'
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensResponse'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult_UsernameIsAlreadyExist_'
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
        - auth
      summary: User Login
      description: Вход пользователя и выдача токенов
      operationId: user_login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult_UserDataIsNotCorrect_'
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/verify:
    security:
      - bearerAuth: []
    get:
      tags:
        - auth
      summary: Token Verify
      description: Проверка токена
      operationId: token_verify_auth_verify_get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyStatus'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult_UserIsNotExists_'
...
  /auth/update_login_data:
    security:
      - bearerAuth: []
    patch:
      tags:
        - auth
      summary: Update User Data
      description: Обнавление данных пользователя
      operationId: update_user_data_auth_users_update_user_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLoginUserRequest'
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateLoginUserResponse'
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Empty:
      title: Empty
      enum:
        - UNSET
      type: string
      description: An enumeration.
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
    ErrorResult_UserDataIsNotCorrect_:
      properties:
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/UserDataIsNotCorrect'
      type: object
      required:
        - message
        - data
      title: ErrorResult[UserDataIsNotCorrect]
      description: 'ErrorResult(*, message: str, data: ~ExcData = None)'
    ErrorResult_UserIsNotExists_:
      properties:
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/UserIsNotExists'
      type: object
      required:
        - message
        - data
      title: ErrorResult[UserIsNotExists]
      description: 'ErrorResult(*, message: str, data: ~ExcData = None)'
    ErrorResult_UsernameIsAlreadyExist_:
      properties:
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/UsernameIsAlreadyExist'
      type: object
      required:
        - message
        - data
      title: ErrorResult[UsernameIsAlreadyExist]
      description: 'ErrorResult(*, message: str, data: ~ExcData = None)'
    LoginRequest:
      properties:
        username:
          type: string
          title: Username
          description: Никнейм пользователя
        password:
          type: string
          title: Password
          description: Пароль пользователя
      type: object
      required:
        - username
        - password
      title: LoginRequest
      description: Модель входа в систему
...
    RegistrationRequest:
      properties:
        username:
          type: string
          title: Username
          description: Никнейм пользователя
        password:
          type: string
          title: Password
          description: Пароль пользователя
        user_email:
          type: string
          format: email
          title: User Email
          description: Почта пользователя
      type: object
      required:
        - username
        - password
        - user_email
      title: RegistrationRequest
      description: Модель регистрации пользоватлея
    TokensResponse:
      properties:
        access_token:
          type: string
          title: Access Token
          description: Jwt для авторизации
        refresh_token:
          type: string
          title: Refresh Token
          description: Рефреш токен для обнавления jwt
        access_token_expires:
          type: integer
          title: Access Token Expires
          description: Время, сколько активен jwt
      type: object
      required:
        - access_token
        - refresh_token
        - access_token_expires
      title: TokensResponse
      description: Выдача токенов
    UpdateLoginUserRequest:
      properties:
        password:
          anyOf:
            - type: string
            - $ref: '#/components/schemas/Empty'
          title: Password
          default: '1'
        user_email:
          anyOf:
            - type: string
              format: email
            - $ref: '#/components/schemas/Empty'
          title: User Email
          default: '1'
      type: object
      title: UpdateUserRequest
      description: Модель для обнавление данных пользователя
    UpdateLoginUserResponse:
      properties:
        username:
          type: string
          title: Username
          description: Никнейм пользователя
      type: object
      required:
        - username
      title: UpdateLoginUserResponse
      description: Модель ответа об обновлении пользователя
    UserDataIsNotCorrect:
      properties: {}
      type: object
      title: UserDataIsNotCorrect
    UserIsNotExists:
      properties:
        user_id:
          type: integer
          title: User Id
      type: object
      required:
        - user_id
      title: UserIsNotExists
    UsernameIsAlreadyExist:
      properties:
        username:
          type: string
          title: Username
      type: object
      required:
        - username
      title: UsernameIsAlreadyExist
    VerifyStatus:
      properties:
        verify_status:
          type: string
          title: Verify Status
          default: Done
      type: object
      title: VerifyStatus
...